Make the rust report more idiomatic
I'm still learning rust and this is the current state of the report that
I'm using in, what I believe to be, a more idiomatic style.

The previous version was more of a direct translation from the python
script to the rust report.

NOTE: `cargo test` currently does not work.